#8 [REFACTOR] API 폴더구조 리팩토링
