+++ src/apis/index.ts
@@ -1,3 +1,2 @@
-export * from './auth';
-export * from './signup';
-export * from './axios';
+export * from "./auth";
+export * from "./axios";
